backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncio>=3.4.3",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "databases>=0.9.0",
    "fastapi[all]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
    "sentry-sdk>=2.24.1",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.2",
]

[tool.pytest.ini_options]
addopts = "-s"
